# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API ASKSUITE",
    "description": "API de consulta a hoteis"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
    "/api/buscar": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "busca",
            "in": "body",
            "schema": {
              "type": "multipart/form-data",
              "properties": {
                "checkin": {
                  "type": "string",
                  "example": "01/01/2021",
                  "required":true
                },
                "checkout": {
                  "type": "string",
                  "example": "01/01/2021",
                  "required":true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso na consulta."
          },
          "400": {
            "description": "Falha na  consulta."
          }
        }
      }
    },
    "/login":{
      "post":{
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "multipart/form-data",
              "properties": {
                "user": {
                  "type": "string",
                  "example": "luiz"
                },
                "password": {
                  "type": "string",
                  "example": "123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso no login."
          },
          "400": {
            "description": "Erro no login."
          }
        }
      }
    },
    "/logout":{
      "post":{
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "Sucesso no logout."
          },
          "400": {
            "description": "Erro no logout."
          }
        }
      }
    }
  },
  "definitions": {}
}